update other functions to use Gaia ID too

+++ src/core/modules/StarMgr.hpp
@@ -256,10 +256,10 @@ public:
 	StelObjectP searchGaia(int64_t source_id) const;
 
 	//! Get the (translated) common name for a star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return translated common name of star
-	static QString getCommonName(int hip);
+	static QString getCommonName(int64_t hip);
 
 	//! Get the (translated) scientific name for a star with a specified
 	//! Hipparcos or Gaia catalogue number.
@@ -268,113 +268,113 @@ public:
 	static QString getSciName(int64_t hip);
 
 	//! Get the (translated) scientific extra name for a star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return translated scientific name of star
-	static QString getSciExtraName(int hip);
+	static QString getSciExtraName(int64_t hip);
 
 	//! Get the (translated) scientific name for a variable star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return translated scientific name of variable star
-	static QString getGcvsName(int hip);
+	static QString getGcvsName(int64_t hip);
 
 	//! Get the (translated) scientific name for a double star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return translated scientific name of double star
-	static QString getWdsName(int hip);
+	static QString getWdsName(int64_t hip);
 
 	//! Get the (English) common name for a star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return common name of star (from skyculture file star_names.fab)
-	static QString getCommonEnglishName(int hip);
+	static QString getCommonEnglishName(int64_t hip);
 
 	//! Get the (translated) additional names for a star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return translated additional names of star
-	static QString getAdditionalNames(int hip);
+	static QString getAdditionalNames(int64_t hip);
 
 	//! Get the English additional names for a star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return additional names of star
-	static QString getAdditionalEnglishNames(int hip);
+	static QString getAdditionalEnglishNames(int64_t hip);
 
 	//! Get the cross-identification designations for a star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return cross-identification data
 	static QString getCrossIdentificationDesignations(QString hip);
 
 	//! Get the type of variability for a variable star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return type of variability
-	static QString getGcvsVariabilityType(int hip);
+	static QString getGcvsVariabilityType(int64_t hip);
 
 	//! Get the magnitude at maximum brightness for a variable star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return the magnitude at maximum brightness for a variable star
-	static float getGcvsMaxMagnitude(int hip);
+	static float getGcvsMaxMagnitude(int64_t hip);
 
 	//! Get the magnitude flag code for a variable star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return the magnitude flag code for a variable star
-	static int getGcvsMagnitudeFlag(int hip);
+	static int getGcvsMagnitudeFlag(int64_t hip);
 
 	//! Get the minimum magnitude or amplitude for a variable star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @param firstMinimumFlag
 	//! @return the minimum magnitude or amplitude for a variable star
-	static float getGcvsMinMagnitude(int hip, bool firstMinimumFlag=true);
+	static float getGcvsMinMagnitude(int64_t hip, bool firstMinimumFlag=true);
 
 	//! Get the photometric system for a variable star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return the photometric system for a variable star
-	static QString getGcvsPhotometricSystem(int hip);
+	static QString getGcvsPhotometricSystem(int64_t hip);
 
 	//! Get Epoch for maximum light for a variable star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return Epoch for maximum light for a variable star
-	static double getGcvsEpoch(int hip);
+	static double getGcvsEpoch(int64_t hip);
 
 	//! Get the period for a variable star with a specified
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return the period of variable star
-	static double getGcvsPeriod(int hip);
+	static double getGcvsPeriod(int64_t hip);
 
 	//! Get the rising time or duration of eclipse for a variable star with a
-	//! specified Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! specified Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return the rising time or duration of eclipse for variable star
-	static int getGcvsMM(int hip);
+	static int getGcvsMM(int64_t hip);
 
 	//! Get year of last satisfactory observation of double star with a
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return year of last satisfactory observation
-	static int getWdsLastObservation(int hip);
+	static int getWdsLastObservation(int64_t hip);
 
 	//! Get position angle at date of last satisfactory observation of double star with a
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return position angle in degrees
-	static float getWdsLastPositionAngle(int hip);
+	static float getWdsLastPositionAngle(int64_t hip);
 
 	//! Get separation angle at date of last satisfactory observation of double star with a
-	//! Hipparcos catalogue number.
-	//! @param hip The Hipparcos number of star
+	//! Hipparcos or Gaia catalogue number.
+	//! @param hip The Hipparcos/Gaia number of star
 	//! @return separation in arcseconds
-	static float getWdsLastSeparation(int hip);
+	static float getWdsLastSeparation(int64_t hip);
 
 	static QString convertToSpectralType(int index);
 	static QString convertToComponentIds(int index);
@@ -495,33 +495,33 @@ private:
 
 	HipIndexStruct *hipIndex; // array of Hipparcos stars
 
-	static QHash<int, QString> commonNamesMap;     // the original names from skyculture (star_names.fab)
-	static QHash<int, QString> commonNamesMapI18n; // translated names
-	static QMap<QString, int> commonNamesIndexI18n;
-	static QMap<QString, int> commonNamesIndex;
+	static QHash<int64_t, QString> commonNamesMap;     // the original names from skyculture (star_names.fab)
+	static QHash<int64_t, QString> commonNamesMapI18n; // translated names
+	static QMap<QString, int64_t> commonNamesIndexI18n;
+	static QMap<QString, int64_t> commonNamesIndex;
 
-	static QHash<int, QString> additionalNamesMap; // additional names
-	static QHash<int, QString> additionalNamesMapI18n;
-	static QMap<QString, int> additionalNamesIndex;
-	static QMap<QString, int> additionalNamesIndexI18n;
+	static QHash<int64_t, QString> additionalNamesMap; // additional names
+	static QHash<int64_t, QString> additionalNamesMapI18n;
+	static QMap<QString, int64_t> additionalNamesIndex;
+	static QMap<QString, int64_t> additionalNamesIndexI18n;
 
 	static QHash<int64_t, QString> sciDesignationsMapI18n;
-	static QMap<QString, int> sciDesignationsIndexI18n;
-	static QHash<int, QString> sciExtraDesignationsMapI18n;
-	static QMap<QString, int> sciExtraDesignationsIndexI18n;
+	static QMap<QString, int64_t> sciDesignationsIndexI18n;
+	static QHash<int64_t, QString> sciExtraDesignationsMapI18n;
+	static QMap<QString, int64_t> sciExtraDesignationsIndexI18n;
 
-	static QHash<int, varstar> varStarsMapI18n;
-	static QMap<QString, int> varStarsIndexI18n;
+	static QHash<int64_t, varstar> varStarsMapI18n;
+	static QMap<QString, int64_t> varStarsIndexI18n;
 
-	static QHash<int, wds> wdsStarsMapI18n;
-	static QMap<QString, int> wdsStarsIndexI18n;
+	static QHash<int64_t, wds> wdsStarsMapI18n;
+	static QMap<QString, int64_t> wdsStarsIndexI18n;
 
 	static QMap<QString, crossid> crossIdMap;
-	static QMap<int, int> saoStarsIndex;
-	static QMap<int, int> hdStarsIndex;
-	static QMap<int, int> hrStarsIndex;
+	static QMap<int64_t, int> saoStarsIndex;
+	static QMap<int64_t, int> hdStarsIndex;
+	static QMap<int64_t, int> hrStarsIndex;
 
-	static QHash<int, QString> referenceMap;
+	static QHash<int64_t, QString> referenceMap;
 
 	QFont starFont;
 	static bool flagSciNames;
